Revert "Game buttons"
